Add files via upload
Based on the use of the Dice game (with the use of random lib).
a GUI has been added. a button to roll dice.
the ability to add or subtract the number of dice rolls has been implemented.
the scope has changed.
the type of dice has been expanded on with the introduction of the concept of "I am making this for DND dice now."
the dice are now tracked in a dnd like fashion "#d#". the first being the count of dice, the second being the number of sides.
7 dice types have been added.
the ability to increase and decrease each die with the mouse scroll wheel has been added.
error handling added.
custom darkmode pop out windows added.
when the #d# of a die is clicked, a pop up window where the person can type in the desired number of dice of that type.