Create UnitConversion.py
Added a GUI.
Added various units to convert to / from.
expanded upon the various types of measurement.
added tabs to swap between the measurement types.
implemented pandas and notebooks to switch the units. 
     (liked this more than the dropdown or radio buttons)
added the scroll wheel function that I have implemented so many times before.
Added a ruler emoji as an icon.